dira_backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
google-genai>=0.3.0
redis>=5.0.0
django-redis>=5.4.0
hiredis>=2.0.0
